Improve cuc test names
The rigid naming format is useful for writing generic stepdefs that match related features, but we also need to be able to easily write different stepdefs for less related features. Putting the equivalent of a tag in the feature name is an easy way to control what the stepdef regex picks up.